Access_Portal/Web_Api/requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
SQLAlchemy==2.0.20
flask-migrate==4.0.4
flask_cors==4.0.0 
bcrypt==4.0.1
PyJWT==2.8.0
APScheduler==3.10.4